+++ java-manta-it/src/test/java/com/joyent/manta/client/multipart/EncryptedJobsMultipartManagerIT.java
@@ -52,7 +52,10 @@ import static org.testng.Assert.assertFalse;
 import static org.testng.Assert.assertTrue;
 import static org.testng.Assert.fail;
 
-@Test(groups = { "encrypted" })
+/* Tests are disabled because functionality is obsolete and overall test runtime
+ * is very long and expensive. */
+@Deprecated
+@Test(groups = { "encrypted" }, enabled = false)
 public class EncryptedJobsMultipartManagerIT {
     private MantaClient mantaClient;
     private EncryptedJobsMultipartManager multipart;